LayoutTests/fast/selectors/querySelector-tagname-case-sensitivity-xml-in-xhtml-expected.txt

Test tagname's case-sensitivity when matching XML-in-XHTML with querySelector.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Basic cases, tagname alone.
PASS document.getElementById('with-renderer').querySelectorAll('rect').length is 3
PASS document.getElementById('without-renderer').querySelectorAll('rect').length is 3
PASS document.getElementById("with-renderer").querySelectorAll('rect')[0].getAttribute('class') is "target1"
PASS document.getElementById("without-renderer").querySelectorAll('rect')[0].getAttribute('class') is "target1"
PASS document.getElementById("with-renderer").querySelectorAll('rect')[1].getAttribute('class') is "target4"
PASS document.getElementById("without-renderer").querySelectorAll('rect')[1].getAttribute('class') is "target4"
PASS document.getElementById("with-renderer").querySelectorAll('rect')[2].getAttribute('class') is "target7"
PASS document.getElementById("without-renderer").querySelectorAll('rect')[2].getAttribute('class') is "target7"
PASS document.getElementById('with-renderer').querySelectorAll('Rect').length is 3
PASS document.getElementById('without-renderer').querySelectorAll('Rect').length is 3
PASS document.getElementById("with-renderer").querySelectorAll('Rect')[0].getAttribute('class') is "target2"
PASS document.getElementById("without-renderer").querySelectorAll('Rect')[0].getAttribute('class') is "target2"
PASS document.getElementById("with-renderer").querySelectorAll('Rect')[1].getAttribute('class') is "target5"
PASS document.getElementById("without-renderer").querySelectorAll('Rect')[1].getAttribute('class') is "target5"
PASS document.getElementById("with-renderer").querySelectorAll('Rect')[2].getAttribute('class') is "target8"
PASS document.getElementById("without-renderer").querySelectorAll('Rect')[2].getAttribute('class') is "target8"
PASS document.getElementById('with-renderer').querySelectorAll('RECT').length is 3
PASS document.getElementById('without-renderer').querySelectorAll('RECT').length is 3
PASS document.getElementById("with-renderer").querySelectorAll('RECT')[0].getAttribute('class') is "target3"
PASS document.getElementById("without-renderer").querySelectorAll('RECT')[0].getAttribute('class') is "target3"
PASS document.getElementById("with-renderer").querySelectorAll('RECT')[1].getAttribute('class') is "target6"
PASS document.getElementById("without-renderer").querySelectorAll('RECT')[1].getAttribute('class') is "target6"
PASS document.getElementById("with-renderer").querySelectorAll('RECT')[2].getAttribute('class') is "target9"
PASS document.getElementById("without-renderer").querySelectorAll('RECT')[2].getAttribute('class') is "target9"

Compound selectors.
PASS document.getElementById('with-renderer').querySelectorAll('rect[class]').length is 3
PASS document.getElementById('without-renderer').querySelectorAll('rect[class]').length is 3
PASS document.getElementById("with-renderer").querySelectorAll('rect[class]')[0].getAttribute('class') is "target1"
PASS document.getElementById("without-renderer").querySelectorAll('rect[class]')[0].getAttribute('class') is "target1"
PASS document.getElementById("with-renderer").querySelectorAll('rect[class]')[1].getAttribute('class') is "target4"
PASS document.getElementById("without-renderer").querySelectorAll('rect[class]')[1].getAttribute('class') is "target4"
PASS document.getElementById("with-renderer").querySelectorAll('rect[class]')[2].getAttribute('class') is "target7"
PASS document.getElementById("without-renderer").querySelectorAll('rect[class]')[2].getAttribute('class') is "target7"
PASS document.getElementById('with-renderer').querySelectorAll('Rect[class]').length is 3
PASS document.getElementById('without-renderer').querySelectorAll('Rect[class]').length is 3
PASS document.getElementById("with-renderer").querySelectorAll('Rect[class]')[0].getAttribute('class') is "target2"
PASS document.getElementById("without-renderer").querySelectorAll('Rect[class]')[0].getAttribute('class') is "target2"
PASS document.getElementById("with-renderer").querySelectorAll('Rect[class]')[1].getAttribute('class') is "target5"
PASS document.getElementById("without-renderer").querySelectorAll('Rect[class]')[1].getAttribute('class') is "target5"
PASS document.getElementById("with-renderer").querySelectorAll('Rect[class]')[2].getAttribute('class') is "target8"
PASS document.getElementById("without-renderer").querySelectorAll('Rect[class]')[2].getAttribute('class') is "target8"
PASS document.getElementById('with-renderer').querySelectorAll('RECT[class]').length is 3
PASS document.getElementById('without-renderer').querySelectorAll('RECT[class]').length is 3
PASS document.getElementById("with-renderer").querySelectorAll('RECT[class]')[0].getAttribute('class') is "target3"
PASS document.getElementById("without-renderer").querySelectorAll('RECT[class]')[0].getAttribute('class') is "target3"
PASS document.getElementById("with-renderer").querySelectorAll('RECT[class]')[1].getAttribute('class') is "target6"
PASS document.getElementById("without-renderer").querySelectorAll('RECT[class]')[1].getAttribute('class') is "target6"
PASS document.getElementById("with-renderer").querySelectorAll('RECT[class]')[2].getAttribute('class') is "target9"
PASS document.getElementById("without-renderer").querySelectorAll('RECT[class]')[2].getAttribute('class') is "target9"
PASS document.getElementById('with-renderer').querySelectorAll('rect[class^="target"]').length is 3
PASS document.getElementById('without-renderer').querySelectorAll('rect[class^="target"]').length is 3
PASS document.getElementById("with-renderer").querySelectorAll('rect[class^="target"]')[0].getAttribute('class') is "target1"
PASS document.getElementById("without-renderer").querySelectorAll('rect[class^="target"]')[0].getAttribute('class') is "target1"
PASS document.getElementById("with-renderer").querySelectorAll('rect[class^="target"]')[1].getAttribute('class') is "target4"
PASS document.getElementById("without-renderer").querySelectorAll('rect[class^="target"]')[1].getAttribute('class') is "target4"
PASS document.getElementById("with-renderer").querySelectorAll('rect[class^="target"]')[2].getAttribute('class') is "target7"
PASS document.getElementById("without-renderer").querySelectorAll('rect[class^="target"]')[2].getAttribute('class') is "target7"
PASS document.getElementById('with-renderer').querySelectorAll('Rect[class^="target"]').length is 3
PASS document.getElementById('without-renderer').querySelectorAll('Rect[class^="target"]').length is 3
PASS document.getElementById("with-renderer").querySelectorAll('Rect[class^="target"]')[0].getAttribute('class') is "target2"
PASS document.getElementById("without-renderer").querySelectorAll('Rect[class^="target"]')[0].getAttribute('class') is "target2"
PASS document.getElementById("with-renderer").querySelectorAll('Rect[class^="target"]')[1].getAttribute('class') is "target5"
PASS document.getElementById("without-renderer").querySelectorAll('Rect[class^="target"]')[1].getAttribute('class') is "target5"
PASS document.getElementById("with-renderer").querySelectorAll('Rect[class^="target"]')[2].getAttribute('class') is "target8"
PASS document.getElementById("without-renderer").querySelectorAll('Rect[class^="target"]')[2].getAttribute('class') is "target8"
PASS document.getElementById('with-renderer').querySelectorAll('RECT[class^="target"]').length is 3
PASS document.getElementById('without-renderer').querySelectorAll('RECT[class^="target"]').length is 3
PASS document.getElementById("with-renderer").querySelectorAll('RECT[class^="target"]')[0].getAttribute('class') is "target3"
PASS document.getElementById("without-renderer").querySelectorAll('RECT[class^="target"]')[0].getAttribute('class') is "target3"
PASS document.getElementById("with-renderer").querySelectorAll('RECT[class^="target"]')[1].getAttribute('class') is "target6"
PASS document.getElementById("without-renderer").querySelectorAll('RECT[class^="target"]')[1].getAttribute('class') is "target6"
PASS document.getElementById("with-renderer").querySelectorAll('RECT[class^="target"]')[2].getAttribute('class') is "target9"
PASS document.getElementById("without-renderer").querySelectorAll('RECT[class^="target"]')[2].getAttribute('class') is "target9"

Complex selectors.
PASS document.getElementById('with-renderer').querySelectorAll('xml > rect[class]').length is 2
PASS document.getElementById('without-renderer').querySelectorAll('xml > rect[class]').length is 2
PASS document.getElementById("with-renderer").querySelectorAll('xml > rect[class]')[0].getAttribute('class') is "target4"
PASS document.getElementById("without-renderer").querySelectorAll('xml > rect[class]')[0].getAttribute('class') is "target4"
PASS document.getElementById("with-renderer").querySelectorAll('xml > rect[class]')[1].getAttribute('class') is "target7"
PASS document.getElementById("without-renderer").querySelectorAll('xml > rect[class]')[1].getAttribute('class') is "target7"
PASS document.getElementById('with-renderer').querySelectorAll('xml > Rect[class]').length is 2
PASS document.getElementById('without-renderer').querySelectorAll('xml > Rect[class]').length is 2
PASS document.getElementById("with-renderer").querySelectorAll('xml > Rect[class]')[0].getAttribute('class') is "target5"
PASS document.getElementById("without-renderer").querySelectorAll('xml > Rect[class]')[0].getAttribute('class') is "target5"
PASS document.getElementById("with-renderer").querySelectorAll('xml > Rect[class]')[1].getAttribute('class') is "target8"
PASS document.getElementById("without-renderer").querySelectorAll('xml > Rect[class]')[1].getAttribute('class') is "target8"
PASS document.getElementById('with-renderer').querySelectorAll('xml > RECT[class]').length is 2
PASS document.getElementById('without-renderer').querySelectorAll('xml > RECT[class]').length is 2
PASS document.getElementById("with-renderer").querySelectorAll('xml > RECT[class]')[0].getAttribute('class') is "target6"
PASS document.getElementById("without-renderer").querySelectorAll('xml > RECT[class]')[0].getAttribute('class') is "target6"
PASS document.getElementById("with-renderer").querySelectorAll('xml > RECT[class]')[1].getAttribute('class') is "target9"
PASS document.getElementById("without-renderer").querySelectorAll('xml > RECT[class]')[1].getAttribute('class') is "target9"
PASS document.getElementById('with-renderer').querySelectorAll('xml > rect[class^="target"]').length is 2
PASS document.getElementById('without-renderer').querySelectorAll('xml > rect[class^="target"]').length is 2
PASS document.getElementById("with-renderer").querySelectorAll('xml > rect[class^="target"]')[0].getAttribute('class') is "target4"
PASS document.getElementById("without-renderer").querySelectorAll('xml > rect[class^="target"]')[0].getAttribute('class') is "target4"
PASS document.getElementById("with-renderer").querySelectorAll('xml > rect[class^="target"]')[1].getAttribute('class') is "target7"
PASS document.getElementById("without-renderer").querySelectorAll('xml > rect[class^="target"]')[1].getAttribute('class') is "target7"
PASS document.getElementById('with-renderer').querySelectorAll('xml > Rect[class^="target"]').length is 2
PASS document.getElementById('without-renderer').querySelectorAll('xml > Rect[class^="target"]').length is 2
PASS document.getElementById("with-renderer").querySelectorAll('xml > Rect[class^="target"]')[0].getAttribute('class') is "target5"
PASS document.getElementById("without-renderer").querySelectorAll('xml > Rect[class^="target"]')[0].getAttribute('class') is "target5"
PASS document.getElementById("with-renderer").querySelectorAll('xml > Rect[class^="target"]')[1].getAttribute('class') is "target8"
PASS document.getElementById("without-renderer").querySelectorAll('xml > Rect[class^="target"]')[1].getAttribute('class') is "target8"
PASS document.getElementById('with-renderer').querySelectorAll('xml > RECT[class^="target"]').length is 2
PASS document.getElementById('without-renderer').querySelectorAll('xml > RECT[class^="target"]').length is 2
PASS document.getElementById("with-renderer").querySelectorAll('xml > RECT[class^="target"]')[0].getAttribute('class') is "target6"
PASS document.getElementById("without-renderer").querySelectorAll('xml > RECT[class^="target"]')[0].getAttribute('class') is "target6"
PASS document.getElementById("with-renderer").querySelectorAll('xml > RECT[class^="target"]')[1].getAttribute('class') is "target9"
PASS document.getElementById("without-renderer").querySelectorAll('xml > RECT[class^="target"]')[1].getAttribute('class') is "target9"
PASS document.getElementById('with-renderer').querySelectorAll('div rect[class]').length is 3
PASS document.getElementById('without-renderer').querySelectorAll('div rect[class]').length is 3
PASS document.getElementById("with-renderer").querySelectorAll('div rect[class]')[0].getAttribute('class') is "target1"
PASS document.getElementById("without-renderer").querySelectorAll('div rect[class]')[0].getAttribute('class') is "target1"
PASS document.getElementById("with-renderer").querySelectorAll('div rect[class]')[1].getAttribute('class') is "target4"
PASS document.getElementById("without-renderer").querySelectorAll('div rect[class]')[1].getAttribute('class') is "target4"
PASS document.getElementById("with-renderer").querySelectorAll('div rect[class]')[2].getAttribute('class') is "target7"
PASS document.getElementById("without-renderer").querySelectorAll('div rect[class]')[2].getAttribute('class') is "target7"
PASS document.getElementById('with-renderer').querySelectorAll('div Rect[class]').length is 3
PASS document.getElementById('without-renderer').querySelectorAll('div Rect[class]').length is 3
PASS document.getElementById("with-renderer").querySelectorAll('div Rect[class]')[0].getAttribute('class') is "target2"
PASS document.getElementById("without-renderer").querySelectorAll('div Rect[class]')[0].getAttribute('class') is "target2"
PASS document.getElementById("with-renderer").querySelectorAll('div Rect[class]')[1].getAttribute('class') is "target5"
PASS document.getElementById("without-renderer").querySelectorAll('div Rect[class]')[1].getAttribute('class') is "target5"
PASS document.getElementById("with-renderer").querySelectorAll('div Rect[class]')[2].getAttribute('class') is "target8"
PASS document.getElementById("without-renderer").querySelectorAll('div Rect[class]')[2].getAttribute('class') is "target8"
PASS document.getElementById('with-renderer').querySelectorAll('div RECT[class]').length is 3
PASS document.getElementById('without-renderer').querySelectorAll('div RECT[class]').length is 3
PASS document.getElementById("with-renderer").querySelectorAll('div RECT[class]')[0].getAttribute('class') is "target3"
PASS document.getElementById("without-renderer").querySelectorAll('div RECT[class]')[0].getAttribute('class') is "target3"
PASS document.getElementById("with-renderer").querySelectorAll('div RECT[class]')[1].getAttribute('class') is "target6"
PASS document.getElementById("without-renderer").querySelectorAll('div RECT[class]')[1].getAttribute('class') is "target6"
PASS document.getElementById("with-renderer").querySelectorAll('div RECT[class]')[2].getAttribute('class') is "target9"
PASS document.getElementById("without-renderer").querySelectorAll('div RECT[class]')[2].getAttribute('class') is "target9"
PASS document.getElementById('with-renderer').querySelectorAll('div rect[class^="target"]').length is 3
PASS document.getElementById('without-renderer').querySelectorAll('div rect[class^="target"]').length is 3
PASS document.getElementById("with-renderer").querySelectorAll('div rect[class^="target"]')[0].getAttribute('class') is "target1"
PASS document.getElementById("without-renderer").querySelectorAll('div rect[class^="target"]')[0].getAttribute('class') is "target1"
PASS document.getElementById("with-renderer").querySelectorAll('div rect[class^="target"]')[1].getAttribute('class') is "target4"
PASS document.getElementById("without-renderer").querySelectorAll('div rect[class^="target"]')[1].getAttribute('class') is "target4"
PASS document.getElementById("with-renderer").querySelectorAll('div rect[class^="target"]')[2].getAttribute('class') is "target7"
PASS document.getElementById("without-renderer").querySelectorAll('div rect[class^="target"]')[2].getAttribute('class') is "target7"
PASS document.getElementById('with-renderer').querySelectorAll('div Rect[class^="target"]').length is 3
PASS document.getElementById('without-renderer').querySelectorAll('div Rect[class^="target"]').length is 3
PASS document.getElementById("with-renderer").querySelectorAll('div Rect[class^="target"]')[0].getAttribute('class') is "target2"
PASS document.getElementById("without-renderer").querySelectorAll('div Rect[class^="target"]')[0].getAttribute('class') is "target2"
PASS document.getElementById("with-renderer").querySelectorAll('div Rect[class^="target"]')[1].getAttribute('class') is "target5"
PASS document.getElementById("without-renderer").querySelectorAll('div Rect[class^="target"]')[1].getAttribute('class') is "target5"
PASS document.getElementById("with-renderer").querySelectorAll('div Rect[class^="target"]')[2].getAttribute('class') is "target8"
PASS document.getElementById("without-renderer").querySelectorAll('div Rect[class^="target"]')[2].getAttribute('class') is "target8"
PASS document.getElementById('with-renderer').querySelectorAll('div RECT[class^="target"]').length is 3
PASS document.getElementById('without-renderer').querySelectorAll('div RECT[class^="target"]').length is 3
PASS document.getElementById("with-renderer").querySelectorAll('div RECT[class^="target"]')[0].getAttribute('class') is "target3"
PASS document.getElementById("without-renderer").querySelectorAll('div RECT[class^="target"]')[0].getAttribute('class') is "target3"
PASS document.getElementById("with-renderer").querySelectorAll('div RECT[class^="target"]')[1].getAttribute('class') is "target6"
PASS document.getElementById("without-renderer").querySelectorAll('div RECT[class^="target"]')[1].getAttribute('class') is "target6"
PASS document.getElementById("with-renderer").querySelectorAll('div RECT[class^="target"]')[2].getAttribute('class') is "target9"
PASS document.getElementById("without-renderer").querySelectorAll('div RECT[class^="target"]')[2].getAttribute('class') is "target9"
PASS document.getElementById('with-renderer').querySelectorAll(':root div rect[class]').length is 3
PASS document.getElementById('without-renderer').querySelectorAll(':root div rect[class]').length is 3
PASS document.getElementById("with-renderer").querySelectorAll(':root div rect[class]')[0].getAttribute('class') is "target1"
PASS document.getElementById("without-renderer").querySelectorAll(':root div rect[class]')[0].getAttribute('class') is "target1"
PASS document.getElementById("with-renderer").querySelectorAll(':root div rect[class]')[1].getAttribute('class') is "target4"
PASS document.getElementById("without-renderer").querySelectorAll(':root div rect[class]')[1].getAttribute('class') is "target4"
PASS document.getElementById("with-renderer").querySelectorAll(':root div rect[class]')[2].getAttribute('class') is "target7"
PASS document.getElementById("without-renderer").querySelectorAll(':root div rect[class]')[2].getAttribute('class') is "target7"
PASS document.getElementById('with-renderer').querySelectorAll(':root div Rect[class]').length is 3
PASS document.getElementById('without-renderer').querySelectorAll(':root div Rect[class]').length is 3
PASS document.getElementById("with-renderer").querySelectorAll(':root div Rect[class]')[0].getAttribute('class') is "target2"
PASS document.getElementById("without-renderer").querySelectorAll(':root div Rect[class]')[0].getAttribute('class') is "target2"
PASS document.getElementById("with-renderer").querySelectorAll(':root div Rect[class]')[1].getAttribute('class') is "target5"
PASS document.getElementById("without-renderer").querySelectorAll(':root div Rect[class]')[1].getAttribute('class') is "target5"
PASS document.getElementById("with-renderer").querySelectorAll(':root div Rect[class]')[2].getAttribute('class') is "target8"
PASS document.getElementById("without-renderer").querySelectorAll(':root div Rect[class]')[2].getAttribute('class') is "target8"
PASS document.getElementById('with-renderer').querySelectorAll(':root div RECT[class]').length is 3
PASS document.getElementById('without-renderer').querySelectorAll(':root div RECT[class]').length is 3
PASS document.getElementById("with-renderer").querySelectorAll(':root div RECT[class]')[0].getAttribute('class') is "target3"
PASS document.getElementById("without-renderer").querySelectorAll(':root div RECT[class]')[0].getAttribute('class') is "target3"
PASS document.getElementById("with-renderer").querySelectorAll(':root div RECT[class]')[1].getAttribute('class') is "target6"
PASS document.getElementById("without-renderer").querySelectorAll(':root div RECT[class]')[1].getAttribute('class') is "target6"
PASS document.getElementById("with-renderer").querySelectorAll(':root div RECT[class]')[2].getAttribute('class') is "target9"
PASS document.getElementById("without-renderer").querySelectorAll(':root div RECT[class]')[2].getAttribute('class') is "target9"
PASS document.getElementById('with-renderer').querySelectorAll(':root div rect[class^="target"]').length is 3
PASS document.getElementById('without-renderer').querySelectorAll(':root div rect[class^="target"]').length is 3
PASS document.getElementById("with-renderer").querySelectorAll(':root div rect[class^="target"]')[0].getAttribute('class') is "target1"
PASS document.getElementById("without-renderer").querySelectorAll(':root div rect[class^="target"]')[0].getAttribute('class') is "target1"
PASS document.getElementById("with-renderer").querySelectorAll(':root div rect[class^="target"]')[1].getAttribute('class') is "target4"
PASS document.getElementById("without-renderer").querySelectorAll(':root div rect[class^="target"]')[1].getAttribute('class') is "target4"
PASS document.getElementById("with-renderer").querySelectorAll(':root div rect[class^="target"]')[2].getAttribute('class') is "target7"
PASS document.getElementById("without-renderer").querySelectorAll(':root div rect[class^="target"]')[2].getAttribute('class') is "target7"
PASS document.getElementById('with-renderer').querySelectorAll(':root div Rect[class^="target"]').length is 3
PASS document.getElementById('without-renderer').querySelectorAll(':root div Rect[class^="target"]').length is 3
PASS document.getElementById("with-renderer").querySelectorAll(':root div Rect[class^="target"]')[0].getAttribute('class') is "target2"
PASS document.getElementById("without-renderer").querySelectorAll(':root div Rect[class^="target"]')[0].getAttribute('class') is "target2"
PASS document.getElementById("with-renderer").querySelectorAll(':root div Rect[class^="target"]')[1].getAttribute('class') is "target5"
PASS document.getElementById("without-renderer").querySelectorAll(':root div Rect[class^="target"]')[1].getAttribute('class') is "target5"
PASS document.getElementById("with-renderer").querySelectorAll(':root div Rect[class^="target"]')[2].getAttribute('class') is "target8"
PASS document.getElementById("without-renderer").querySelectorAll(':root div Rect[class^="target"]')[2].getAttribute('class') is "target8"
PASS document.getElementById('with-renderer').querySelectorAll(':root div RECT[class^="target"]').length is 3
PASS document.getElementById('without-renderer').querySelectorAll(':root div RECT[class^="target"]').length is 3
PASS document.getElementById("with-renderer").querySelectorAll(':root div RECT[class^="target"]')[0].getAttribute('class') is "target3"
PASS document.getElementById("without-renderer").querySelectorAll(':root div RECT[class^="target"]')[0].getAttribute('class') is "target3"
PASS document.getElementById("with-renderer").querySelectorAll(':root div RECT[class^="target"]')[1].getAttribute('class') is "target6"
PASS document.getElementById("without-renderer").querySelectorAll(':root div RECT[class^="target"]')[1].getAttribute('class') is "target6"
PASS document.getElementById("with-renderer").querySelectorAll(':root div RECT[class^="target"]')[2].getAttribute('class') is "target9"
PASS document.getElementById("without-renderer").querySelectorAll(':root div RECT[class^="target"]')[2].getAttribute('class') is "target9"

With functional pseudo classes.
PASS document.getElementById('with-renderer').querySelectorAll(':root div :matches(not-there, rect[class])').length is 3
PASS document.getElementById('without-renderer').querySelectorAll(':root div :matches(not-there, rect[class])').length is 3
PASS document.getElementById("with-renderer").querySelectorAll(':root div :matches(not-there, rect[class])')[0].getAttribute('class') is "target1"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :matches(not-there, rect[class])')[0].getAttribute('class') is "target1"
PASS document.getElementById("with-renderer").querySelectorAll(':root div :matches(not-there, rect[class])')[1].getAttribute('class') is "target4"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :matches(not-there, rect[class])')[1].getAttribute('class') is "target4"
PASS document.getElementById("with-renderer").querySelectorAll(':root div :matches(not-there, rect[class])')[2].getAttribute('class') is "target7"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :matches(not-there, rect[class])')[2].getAttribute('class') is "target7"
PASS document.getElementById('with-renderer').querySelectorAll(':root div :matches(not-there, Rect[class])').length is 3
PASS document.getElementById('without-renderer').querySelectorAll(':root div :matches(not-there, Rect[class])').length is 3
PASS document.getElementById("with-renderer").querySelectorAll(':root div :matches(not-there, Rect[class])')[0].getAttribute('class') is "target2"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :matches(not-there, Rect[class])')[0].getAttribute('class') is "target2"
PASS document.getElementById("with-renderer").querySelectorAll(':root div :matches(not-there, Rect[class])')[1].getAttribute('class') is "target5"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :matches(not-there, Rect[class])')[1].getAttribute('class') is "target5"
PASS document.getElementById("with-renderer").querySelectorAll(':root div :matches(not-there, Rect[class])')[2].getAttribute('class') is "target8"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :matches(not-there, Rect[class])')[2].getAttribute('class') is "target8"
PASS document.getElementById('with-renderer').querySelectorAll(':root div :matches(not-there, RECT[class])').length is 3
PASS document.getElementById('without-renderer').querySelectorAll(':root div :matches(not-there, RECT[class])').length is 3
PASS document.getElementById("with-renderer").querySelectorAll(':root div :matches(not-there, RECT[class])')[0].getAttribute('class') is "target3"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :matches(not-there, RECT[class])')[0].getAttribute('class') is "target3"
PASS document.getElementById("with-renderer").querySelectorAll(':root div :matches(not-there, RECT[class])')[1].getAttribute('class') is "target6"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :matches(not-there, RECT[class])')[1].getAttribute('class') is "target6"
PASS document.getElementById("with-renderer").querySelectorAll(':root div :matches(not-there, RECT[class])')[2].getAttribute('class') is "target9"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :matches(not-there, RECT[class])')[2].getAttribute('class') is "target9"
PASS document.getElementById('with-renderer').querySelectorAll(':root div :matches(not-there, rect[class^="target"])').length is 3
PASS document.getElementById('without-renderer').querySelectorAll(':root div :matches(not-there, rect[class^="target"])').length is 3
PASS document.getElementById("with-renderer").querySelectorAll(':root div :matches(not-there, rect[class^="target"])')[0].getAttribute('class') is "target1"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :matches(not-there, rect[class^="target"])')[0].getAttribute('class') is "target1"
PASS document.getElementById("with-renderer").querySelectorAll(':root div :matches(not-there, rect[class^="target"])')[1].getAttribute('class') is "target4"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :matches(not-there, rect[class^="target"])')[1].getAttribute('class') is "target4"
PASS document.getElementById("with-renderer").querySelectorAll(':root div :matches(not-there, rect[class^="target"])')[2].getAttribute('class') is "target7"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :matches(not-there, rect[class^="target"])')[2].getAttribute('class') is "target7"
PASS document.getElementById('with-renderer').querySelectorAll(':root div :matches(not-there, Rect[class^="target"])').length is 3
PASS document.getElementById('without-renderer').querySelectorAll(':root div :matches(not-there, Rect[class^="target"])').length is 3
PASS document.getElementById("with-renderer").querySelectorAll(':root div :matches(not-there, Rect[class^="target"])')[0].getAttribute('class') is "target2"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :matches(not-there, Rect[class^="target"])')[0].getAttribute('class') is "target2"
PASS document.getElementById("with-renderer").querySelectorAll(':root div :matches(not-there, Rect[class^="target"])')[1].getAttribute('class') is "target5"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :matches(not-there, Rect[class^="target"])')[1].getAttribute('class') is "target5"
PASS document.getElementById("with-renderer").querySelectorAll(':root div :matches(not-there, Rect[class^="target"])')[2].getAttribute('class') is "target8"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :matches(not-there, Rect[class^="target"])')[2].getAttribute('class') is "target8"
PASS document.getElementById('with-renderer').querySelectorAll(':root div :matches(not-there, RECT[class^="target"])').length is 3
PASS document.getElementById('without-renderer').querySelectorAll(':root div :matches(not-there, RECT[class^="target"])').length is 3
PASS document.getElementById("with-renderer").querySelectorAll(':root div :matches(not-there, RECT[class^="target"])')[0].getAttribute('class') is "target3"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :matches(not-there, RECT[class^="target"])')[0].getAttribute('class') is "target3"
PASS document.getElementById("with-renderer").querySelectorAll(':root div :matches(not-there, RECT[class^="target"])')[1].getAttribute('class') is "target6"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :matches(not-there, RECT[class^="target"])')[1].getAttribute('class') is "target6"
PASS document.getElementById("with-renderer").querySelectorAll(':root div :matches(not-there, RECT[class^="target"])')[2].getAttribute('class') is "target9"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :matches(not-there, RECT[class^="target"])')[2].getAttribute('class') is "target9"
PASS document.getElementById('with-renderer').querySelectorAll(':root div :nth-child(n of not-there, rect[class])').length is 3
PASS document.getElementById('without-renderer').querySelectorAll(':root div :nth-child(n of not-there, rect[class])').length is 3
PASS document.getElementById("with-renderer").querySelectorAll(':root div :nth-child(n of not-there, rect[class])')[0].getAttribute('class') is "target1"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :nth-child(n of not-there, rect[class])')[0].getAttribute('class') is "target1"
PASS document.getElementById("with-renderer").querySelectorAll(':root div :nth-child(n of not-there, rect[class])')[1].getAttribute('class') is "target4"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :nth-child(n of not-there, rect[class])')[1].getAttribute('class') is "target4"
PASS document.getElementById("with-renderer").querySelectorAll(':root div :nth-child(n of not-there, rect[class])')[2].getAttribute('class') is "target7"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :nth-child(n of not-there, rect[class])')[2].getAttribute('class') is "target7"
PASS document.getElementById('with-renderer').querySelectorAll(':root div :nth-child(n of not-there, Rect[class])').length is 3
PASS document.getElementById('without-renderer').querySelectorAll(':root div :nth-child(n of not-there, Rect[class])').length is 3
PASS document.getElementById("with-renderer").querySelectorAll(':root div :nth-child(n of not-there, Rect[class])')[0].getAttribute('class') is "target2"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :nth-child(n of not-there, Rect[class])')[0].getAttribute('class') is "target2"
PASS document.getElementById("with-renderer").querySelectorAll(':root div :nth-child(n of not-there, Rect[class])')[1].getAttribute('class') is "target5"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :nth-child(n of not-there, Rect[class])')[1].getAttribute('class') is "target5"
PASS document.getElementById("with-renderer").querySelectorAll(':root div :nth-child(n of not-there, Rect[class])')[2].getAttribute('class') is "target8"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :nth-child(n of not-there, Rect[class])')[2].getAttribute('class') is "target8"
PASS document.getElementById('with-renderer').querySelectorAll(':root div :nth-child(n of not-there, RECT[class])').length is 3
PASS document.getElementById('without-renderer').querySelectorAll(':root div :nth-child(n of not-there, RECT[class])').length is 3
PASS document.getElementById("with-renderer").querySelectorAll(':root div :nth-child(n of not-there, RECT[class])')[0].getAttribute('class') is "target3"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :nth-child(n of not-there, RECT[class])')[0].getAttribute('class') is "target3"
PASS document.getElementById("with-renderer").querySelectorAll(':root div :nth-child(n of not-there, RECT[class])')[1].getAttribute('class') is "target6"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :nth-child(n of not-there, RECT[class])')[1].getAttribute('class') is "target6"
PASS document.getElementById("with-renderer").querySelectorAll(':root div :nth-child(n of not-there, RECT[class])')[2].getAttribute('class') is "target9"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :nth-child(n of not-there, RECT[class])')[2].getAttribute('class') is "target9"
PASS document.getElementById('with-renderer').querySelectorAll(':root div :nth-child(n of not-there, rect[class^="target"])').length is 3
PASS document.getElementById('without-renderer').querySelectorAll(':root div :nth-child(n of not-there, rect[class^="target"])').length is 3
PASS document.getElementById("with-renderer").querySelectorAll(':root div :nth-child(n of not-there, rect[class^="target"])')[0].getAttribute('class') is "target1"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :nth-child(n of not-there, rect[class^="target"])')[0].getAttribute('class') is "target1"
PASS document.getElementById("with-renderer").querySelectorAll(':root div :nth-child(n of not-there, rect[class^="target"])')[1].getAttribute('class') is "target4"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :nth-child(n of not-there, rect[class^="target"])')[1].getAttribute('class') is "target4"
PASS document.getElementById("with-renderer").querySelectorAll(':root div :nth-child(n of not-there, rect[class^="target"])')[2].getAttribute('class') is "target7"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :nth-child(n of not-there, rect[class^="target"])')[2].getAttribute('class') is "target7"
PASS document.getElementById('with-renderer').querySelectorAll(':root div :nth-child(n of not-there, Rect[class^="target"])').length is 3
PASS document.getElementById('without-renderer').querySelectorAll(':root div :nth-child(n of not-there, Rect[class^="target"])').length is 3
PASS document.getElementById("with-renderer").querySelectorAll(':root div :nth-child(n of not-there, Rect[class^="target"])')[0].getAttribute('class') is "target2"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :nth-child(n of not-there, Rect[class^="target"])')[0].getAttribute('class') is "target2"
PASS document.getElementById("with-renderer").querySelectorAll(':root div :nth-child(n of not-there, Rect[class^="target"])')[1].getAttribute('class') is "target5"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :nth-child(n of not-there, Rect[class^="target"])')[1].getAttribute('class') is "target5"
PASS document.getElementById("with-renderer").querySelectorAll(':root div :nth-child(n of not-there, Rect[class^="target"])')[2].getAttribute('class') is "target8"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :nth-child(n of not-there, Rect[class^="target"])')[2].getAttribute('class') is "target8"
PASS document.getElementById('with-renderer').querySelectorAll(':root div :nth-child(n of not-there, RECT[class^="target"])').length is 3
PASS document.getElementById('without-renderer').querySelectorAll(':root div :nth-child(n of not-there, RECT[class^="target"])').length is 3
PASS document.getElementById("with-renderer").querySelectorAll(':root div :nth-child(n of not-there, RECT[class^="target"])')[0].getAttribute('class') is "target3"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :nth-child(n of not-there, RECT[class^="target"])')[0].getAttribute('class') is "target3"
PASS document.getElementById("with-renderer").querySelectorAll(':root div :nth-child(n of not-there, RECT[class^="target"])')[1].getAttribute('class') is "target6"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :nth-child(n of not-there, RECT[class^="target"])')[1].getAttribute('class') is "target6"
PASS document.getElementById("with-renderer").querySelectorAll(':root div :nth-child(n of not-there, RECT[class^="target"])')[2].getAttribute('class') is "target9"
PASS document.getElementById("without-renderer").querySelectorAll(':root div :nth-child(n of not-there, RECT[class^="target"])')[2].getAttribute('class') is "target9"
PASS successfullyParsed is true

TEST COMPLETE
